skills/ketchup-code-viz/requirements.txt
# Ketchup Code Viz Dependencies
Pillow>=10.0.0
pygments>=2.16.0
GitPython>=3.1.0  # Optional, for commit extraction
